Moved CEScrollView to separate file and used documentRange.endLocation to simplify textView range calculation (#56)
